requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.2
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.4
pillow==11.3.0
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
